requirements.txt
click==7.1.2
python-dotenv==0.15.0
requests==2.25.1
